Updated the source code for PSF_Control_Terminal and the respective files as per the new EVB-UNG_8270

+++ PSF/Demo/PSF_EVB_Sink/firmware/src/config/default/peripheral/sercom/spim/plib_sercom0_spi.c
@@ -159,7 +159,7 @@ bool SERCOM0_SPI_TransferSetup(SPI_TRANSFER_SETUP *setup, uint32_t spiSourceCloc
         if((baudValue > 0) & (baudValue <= 255))
         {
             /* Selection of the Clock Polarity and Clock Phase */
-            SERCOM0_REGS->SPIM.SERCOM_CTRLA |= setup->clockPolarity | setup->clockPhase;
+            SERCOM0_REGS->SPIM.SERCOM_CTRLA |= (uint32_t)setup->clockPolarity | (uint32_t)setup->clockPhase;
 
             /* Selection of the Baud Value */
             SERCOM0_REGS->SPIM.SERCOM_BAUD = baudValue;
